refactor - merged challenge_exit_test into invalid_exit_test
for faster testing - this is the same pipeline
